pyproject.toml
[project]
name = "agents-observation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonpatch>=1.33",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.1",
    "tzdata>=2025.3",
    # OpenTelemetry SDK + OTLP export + GenAI (OpenAI) instrumentation
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",
    "opentelemetry-instrumentation-openai-v2>=2.1b0",
]
